[DS,ML] Move Adapter classes to adapter package
