+++ app/src/main/java/org/catroid/catrobat/newui/recycleviewlist/adapter/RecyclerViewAdapterDelegate.java
@@ -1,4 +1,4 @@
-package org.catroid.catrobat.newui.recycleviewlist;
+package org.catroid.catrobat.newui.recycleviewlist.adapter;
 
 import org.catroid.catrobat.newui.recycleviewlist.adapter.RecyclerViewAdapter;
